chore(docker): Fix LD_LIBRARY_PATH for dockerw
For the basic docker image, LD_LIBRARY_PATH was not properly set and so
native tools were linking against some libraries they shouldn't have
been in /usr/ lib paths. This didn't really manifest as a problem except
that cmake kept linking against our libcurl that we compile in the
Dockerfile and then threw a fit about version information. This didn't
happen in devcontainer because it properly exports LD_LIBRARY_PATH.